+++ c7n/filters/related.py
@@ -111,3 +111,65 @@ class RelatedResourceFilter(ValueFilter):
     def process(self, resources, event=None):
         related = self.get_related(resources)
         return [r for r in resources if self.process_resource(r, related)]
+
+
+class RelatedResourceByIdFilter(RelatedResourceFilter):
+    """
+    Value filter for related resources in which the main resource only contains the related
+    resource id.
+    """
+
+    RelatedResourceByIdExpression = None
+
+    def get_related(self, resources):
+        resource_manager = self.get_resource_manager()
+        related_ids = self.get_related_ids(resources)
+
+        related = {}
+        for r in resource_manager.resources():
+            matched_vpc = self.get_related_by_ids(r) & related_ids
+            if matched_vpc:
+                for vpc in matched_vpc:
+                    related_resources = related.get(vpc, [])
+                    related_resources.append(r)
+                    related[vpc] = related_resources
+        return related
+
+    def get_related_by_ids(self, resources):
+        RelatedResourceKey = self.RelatedResourceByIdExpression or self.RelatedIdsExpression
+        ids = jmespath.search("%s" % RelatedResourceKey, resources)
+        if isinstance(ids, str):
+            ids = [ids]
+        return set(ids)
+
+    def process_resource(self, resource, related):
+        related_ids = self.get_related_ids([resource])
+        op = self.data.get('operator', 'or')
+        found = []
+
+        if self.data.get('match-resource') is True:
+            self.data['value'] = self.get_resource_value(
+                self.data['key'], resource)
+
+        if self.data.get('value_type') == 'resource_count':
+            count_matches = OPERATORS[self.data.get('op')](len(related_ids), self.data.get('value'))
+            if count_matches:
+                self._add_annotations(related_ids, resource)
+            return count_matches
+
+        for rid in related_ids:
+            robjs = related.get(rid, [None])
+            for robj in robjs:
+                if robj is None:
+                    continue
+                if self.match(robj):
+                    found.append(rid)
+
+        if found:
+            self._add_annotations(found, resource)
+
+        if op == 'or' and found:
+            return True
+        elif op == 'and' and len(found) == len(related_ids):
+            return True
+        return False
